Tests/LibWeb/Text/expected/WebAnimations/animation-properties/finished.txt

finished promise remains after finishing: true
finished promise updates after playing: true
cancel() updates finished promise: true
Expected finished promise cancellation
